Add base interfaces for command & query, restructure packages according to library's public api
