--- commandbus/src/main/java/net/dathoang/cqrs/commandbus/Message.java
@@ -1,4 +1,0 @@
-package net.dathoang.cqrs.commandbus;
-
-public interface Message<R> {
-}
