--- commandbus/src/main/java/net/dathoang/cqrs/commandbus/MessageBus.java
@@ -1,5 +1,0 @@
-package net.dathoang.cqrs.commandbus;
-
-interface MessageBus {
-  <R> R dispatch(Message<R> message) throws Exception;
-}
