--- commandbus/src/main/java/net/dathoang/cqrs/commandbus/MessageHandler.java
@@ -1,5 +1,0 @@
-package net.dathoang.cqrs.commandbus;
-
-interface MessageHandler<T extends Message<R>, R> {
-  R handle(T message) throws Exception;
-}
