+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/command/Command.java
@@ -1,0 +1,6 @@
+package net.dathoang.cqrs.commandbus.command;
+
+import net.dathoang.cqrs.commandbus.message.Message;
+
+public interface Command<R> extends Message<R> {
+}
