+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/command/CommandBus.java
@@ -1,0 +1,5 @@
+package net.dathoang.cqrs.commandbus.command;
+
+public interface CommandBus {
+  <R> R dispatch(Command<R> command) throws Exception;
+}
