+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/command/CommandHandler.java
@@ -1,0 +1,5 @@
+package net.dathoang.cqrs.commandbus.command;
+
+public interface CommandHandler<C extends Command<R>, R> {
+  R handle(C command) throws Exception;
+}
