+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/command/CommandHandlerFactory.java
@@ -1,0 +1,5 @@
+package net.dathoang.cqrs.commandbus.command;
+
+public interface CommandHandlerFactory {
+  <R> CommandHandler<Command<R>, R> createHandler(String commandName);
+}
