+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/exceptions/NoHandlerFoundException.java
@@ -1,6 +1,6 @@
 package net.dathoang.cqrs.commandbus.exceptions;
 
-import net.dathoang.cqrs.commandbus.Message;
+import net.dathoang.cqrs.commandbus.message.Message;
 
 public class NoHandlerFoundException extends RuntimeException {
   public NoHandlerFoundException(Class<? extends Message> messageClass) {
