+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/factory/MessageBus.java
@@ -1,0 +1,7 @@
+package net.dathoang.cqrs.commandbus.factory;
+
+import net.dathoang.cqrs.commandbus.message.Message;
+
+interface MessageBus {
+  <R> R dispatch(Message<R> message) throws Exception;
+}
