+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/factory/MessageHandler.java
@@ -1,0 +1,7 @@
+package net.dathoang.cqrs.commandbus.factory;
+
+import net.dathoang.cqrs.commandbus.message.Message;
+
+interface MessageHandler<T extends Message<R>, R> {
+  R handle(T message) throws Exception;
+}
