+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/factory/MessageHandlerFactory.java
@@ -1,4 +1,6 @@
-package net.dathoang.cqrs.commandbus;
+package net.dathoang.cqrs.commandbus.factory;
+
+import net.dathoang.cqrs.commandbus.message.Message;
 
 interface MessageHandlerFactory {
   <R> MessageHandler<Message<R>, R> createHandler(String messageName);
