+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/message/Message.java
@@ -1,0 +1,4 @@
+package net.dathoang.cqrs.commandbus.message;
+
+public interface Message<R> {
+}
