+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/middleware/Middleware.java
@@ -1,4 +1,6 @@
-package net.dathoang.cqrs.commandbus;
+package net.dathoang.cqrs.commandbus.middleware;
+
+import net.dathoang.cqrs.commandbus.message.Message;
 
 public interface Middleware {
   <R> void preHandle(Message<R> message, ResultAndExceptionHolder<R> resultAndExceptionHolder);
