+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/query/Query.java
@@ -1,0 +1,6 @@
+package net.dathoang.cqrs.commandbus.query;
+
+import net.dathoang.cqrs.commandbus.message.Message;
+
+public interface Query<R> extends Message<R> {
+}
