+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/query/QueryBus.java
@@ -1,0 +1,5 @@
+package net.dathoang.cqrs.commandbus.query;
+
+public interface QueryBus {
+  <R> R dispatch(Query<R> query) throws Exception;
+}
