+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/query/QueryHandler.java
@@ -1,0 +1,5 @@
+package net.dathoang.cqrs.commandbus.query;
+
+interface QueryHandler<Q extends Query<R>, R> {
+  R handle(Q query) throws Exception;
+}
