+++ commandbus/src/main/java/net/dathoang/cqrs/commandbus/query/QueryHandlerFactory.java
@@ -1,0 +1,5 @@
+package net.dathoang.cqrs.commandbus.query;
+
+public interface QueryHandlerFactory {
+  <R> QueryHandler<Query<R>, R> createHandler(String queryName);
+}
